namespace + file directory refactoring
